requirements.txt
# Core dependencies for Football League Standings Dashboard
pandas>=1.5.0
requests>=2.28.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
mysql-connector-python>=8.0.0

# Jupyter notebook dependencies
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
